add candidate detail
