+++ lib/features/candidates/candidate_detail/cubit/candidate_detail_cubit.dart
@@ -1,0 +1,8 @@
+import 'package:bloc/bloc.dart';
+import 'package:meta/meta.dart';
+
+part 'candidate_detail_state.dart';
+
+class CandidateDetailCubit extends Cubit<CandidateDetailState> {
+  CandidateDetailCubit() : super(CandidateDetailInitial());
+}
