+++ lib/features/candidates/candidate_detail/cubit/candidate_detail_state.dart
@@ -1,0 +1,6 @@
+part of 'candidate_detail_cubit.dart';
+
+@immutable
+abstract class CandidateDetailState {}
+
+class CandidateDetailInitial extends CandidateDetailState {}
